Create factorial of number
The code begins by taking an integer input `n` from the user. It then initializes a variable `fact` to 1, which will be used to store the computed factorial. Using a `for` loop, the program iterates from `n` down to 1, multiplying the current value of `fact` by the loop variable `i` during each iteration. After the loop completes, `fact` contains the factorial of `n`. Finally, the program prints the computed factorial to the console.